Move kron_matvec to utils
