+++ pfb/utils/utils.py
@@ -9,6 +9,19 @@ from ducc0.fft import r2c, c2r
 iFs = np.fft.ifftshift
 Fs = np.fft.fftshift
 
+def kron_matvec(A, b):
+    D = len(A)
+    N = b.size
+    x = b
+
+    for d in range(D):
+        Gd = A[d].shape[0]
+        NGd = N//Gd
+        X = np.reshape(x, (Gd, NGd))
+        Z = A[d].dot(X).T
+        x = Z.ravel()
+    return x
+
 def prox_21(v, sigma, weights, axis=1):
     """
     Computes weighted version of
